Fixing some deployment issues, tidying up page
