+++ packages/app/.env.dist
@@ -2,3 +2,5 @@ NG_APP_ADRIA_API_URL=http://localhost:3333
 NG_APP_WEB_API_URL=http://localhost:5000
 NG_APP_SPLASH_APP_NAME=ReefGuide
 NG_APP_SPLASH_ADMIN_EMAIL=admin@email.com
+NG_APP_DOCUMENTATION_LINK=https://github.com/open-AIMS/reefguide
+NG_APP_ABOUT_LINK=https://github.com/open-AIMS/reefguide
